+++ src/httplib_config_options.c
@@ -64,13 +64,9 @@ struct httplib_option XX_httplib_config_options[] = {
 	{ "ssl_cipher_list",             CONFIG_TYPE_STRING,      NULL                                                               },
 	{ "ssl_protocol_version",        CONFIG_TYPE_NUMBER,      "0"                                                                },
 	{ "ssl_short_trust",             CONFIG_TYPE_BOOLEAN,     "no"                                                               },
-#if defined(USE_WEBSOCKET)
 	{ "websocket_timeout_ms",        CONFIG_TYPE_NUMBER,      "30000"                                                            },
-#endif
 	{ "decode_url",                  CONFIG_TYPE_BOOLEAN,     "yes"                                                              },
-#if defined(USE_WEBSOCKET)
 	{ "websocket_root",              CONFIG_TYPE_DIRECTORY,   NULL                                                               },
-#endif
 	{ "access_control_allow_origin", CONFIG_TYPE_STRING,      "*"                                                                },
 	{ "error_pages",                 CONFIG_TYPE_DIRECTORY,   NULL                                                               },
 	{ "tcp_nodelay",                 CONFIG_TYPE_NUMBER,      "0"                                                                },